+++ src/nix/mod.rs
@@ -1,4 +1,5 @@
 pub mod devour_flake;
 pub mod eval;
 pub mod lock;
+pub mod url;
 pub mod util;
